requirements.txt
pandas
sentence-transformers
faiss-cpu
transformers
scikit-learn
torch
